add bolus tone and int16 for delay

+++ Firmware/SyringeDriverHW10_Keil/Core/Inc/user_syringe.h
@@ -21,7 +21,7 @@
 
 extern uint16_t syringeTypes[];
 extern uint16_t syringeTimes[];
-extern uint8_t syringWakeups[][2];
+extern uint16_t syringWakeups[][2];
 extern uint8_t systemError;
 extern char *systemErrorMsg[];
 extern uint8_t delayAlarwakups[];
